requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
colorama==0.4.3
Django==3.0.2
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
